+++ java/src/main/java/io/cucumber/messages/ndjson/Jackson2.java
@@ -1,32 +1,99 @@
 package io.cucumber.messages.ndjson;
 
-import java.util.List;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.MapperFeature;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.databind.cfg.ConstructorDetector;
+import com.fasterxml.jackson.databind.introspect.AnnotatedMember;
+import com.fasterxml.jackson.databind.introspect.NopAnnotationIntrospector;
+import com.fasterxml.jackson.databind.json.JsonMapper;
+import com.fasterxml.jackson.databind.module.SimpleModule;
+import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
+import io.cucumber.messages.Property;
+import org.jspecify.annotations.Nullable;
 
-public final class Jackson2 extends AbstractJsonFactory {
+import java.io.IOException;
+import java.io.Serial;
+
+import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_ABSENT;
+
+public final class Jackson2 implements Json {
+
+    private final JsonMapper delegate = JsonMapper.builder()
+            .addModule(new Jdk8Module())
+            .addModule(new CucumberParameterNamesModule())
+            .defaultPropertyInclusion(JsonInclude.Value.construct(NON_ABSENT, NON_ABSENT))
+            .constructorDetector(ConstructorDetector.USE_PROPERTIES_BASED)
+            .enable(SerializationFeature.WRITE_ENUMS_USING_TO_STRING)
+            .enable(DeserializationFeature.READ_ENUMS_USING_TO_STRING)
+            .enable(DeserializationFeature.USE_LONG_FOR_INTS)
+            .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
+            .disable(JsonGenerator.Feature.AUTO_CLOSE_TARGET)
+            .disable(MapperFeature.CAN_OVERRIDE_ACCESS_MODIFIERS)
+            .build();
 
     @Override
-    protected <T> Deserializer<T> createDeserializer(Class<T> type) {
-        return new Jackson2Json<>(type);
+    public <T> Deserializer<T> deserializer(Class<T> type) {
+        return json -> {
+            try {
+                return delegate.readValue(json, type);
+            } catch (JsonProcessingException e) {
+                throw new RuntimeException(e);
+            }
+        };
     }
 
     @Override
-    protected <T> Serializer<T> createSerializer(Class<T> type) {
-        return new Jackson2Json<>(type);
+    public <T> Serializer<T> serializer(Class<T> type) {
+        return (writer, value) -> {
+            try {
+                delegate.writeValue(writer, value);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        };
     }
 
-    @Override
-    List<Dependency> dependencies() {
-        return List.of(
-                new Dependency(
-                        "com.fasterxml.jackson.databind.json.JsonMapper",
-                        "com.fasterxml.jackson.core",
-                        "jackson-databind"
-                ),
-                new Dependency(
-                        "com.fasterxml.jackson.datatype.jdk8.Jdk8Module",
-                        "com.fasterxml.jackson.datatype",
-                        "jackson-datatype-jdk8"
-                )
-        );
+    static final class CucumberParameterNamesModule extends SimpleModule {
+        @Serial
+        private static final long serialVersionUID = 1L;
+
+        CucumberParameterNamesModule() {
+            super("CucumberParameterNamesModule");
+        }
+
+        @Override
+        public void setupModule(SetupContext context) {
+            super.setupModule(context);
+            context.insertAnnotationIntrospector(new PropertyNameAnnotationIntrospector());
+        }
+
+        @Override
+        public int hashCode() {
+            return getClass().hashCode();
+        }
+
+        @Override
+        public boolean equals(Object o) {
+            return this == o;
+        }
+
+        static final class PropertyNameAnnotationIntrospector extends NopAnnotationIntrospector {
+            @Serial
+            private static final long serialVersionUID = 1L;
+
+            @Override
+            public @Nullable String findImplicitPropertyName(AnnotatedMember m) {
+                Property annotation = m.getAnnotation(Property.class);
+                if (annotation == null) {
+                    return null;
+                }
+                return annotation.value();
+            }
+
+        }
     }
 }
