+++ java/src/main/java/io/cucumber/messages/ndjson/Json.java
@@ -1,15 +1,9 @@
 package io.cucumber.messages.ndjson;
 
-public interface JsonFactory {
+public interface Json {
 
     <T> Deserializer<T> deserializer(Class<T> type);
 
     <T> Serializer<T> serializer(Class<T> type);
-
-    boolean enabled();
-
-    static JsonFactory instance() {
-        return JsonFactoryInstance.instance();
-    }
 
 }
