+++ java/src/main/java/io/cucumber/messages/ndjson/JsonProvider.java
@@ -1,0 +1,16 @@
+package io.cucumber.messages.ndjson;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface JsonProvider {
+
+    Optional<Json> instance();
+
+    List<Dependency> dependencies();
+
+    record Dependency(String className, String groupId, String artifactId) {
+
+    }
+
+}
