+++ java/src/main/resources/io.cucumber.messages.ndjson.JsonProvider
@@ -1,0 +1,2 @@
+io.cucumber.messages.ndjson.Jackson2JsonProvider
+io.cucumber.messages.ndjson.Jackson3JsonProvider
